+++ CHANGELOG.md
@@ -3,7 +3,7 @@ See also [the changelog of monaco-editor](https://github.com/Microsoft/monaco-ed
 # 0.19.0
 
 * Add a new widget: `monacoEditorFramed`. It load the monaco editor in an iframe for improved encapsulation.
-* Update to monaco editor `0.19.0`
+* Update to monaco editor `0.19.3`
 
 # 0.18.1